+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - changed: Wording in light account persistent notification
 - fixed: Fix error massaging in trackError
 - fixed: Normalized error messages for tracking; removing localization from error messages.
+- fixed: Sentry tagging and metadata data wasn't formed properly and caused some loss of tracked errors.
 - fixed: Use Sentry context for logging metadata in `EdgeCrashEvent`
 - removed: Bank Wire Transfer Buy for Florida
 - removed: Paypal Sell for Canada